requirements.txt
asgiref==3.10.0
dataclasses==0.6
Django==5.2.7
dnspython==2.8.0
python-decouple==3.8
sqlparse==0.5.3
psycopg2-binary
celery
redis
django-ckeditor
pillow
django-modeltranslationr